fix(vision): 目标选择挪出 Windows 专用类，测试随之跨平台
CI 在 linux 与 macOS 上失败三条，本地全绿掩盖了它：PickTarget 原来接 WindowsWindowService，
而 EnumerateTopLevelWindows 开头就判 Availability，非 Windows 直接返回空列表，用例却期望
选出窗口。

修法不是给用例加平台守卫 —— 那只会让它们在两个平台上静默跳过，而「不能截到自己」恰恰是
这里最容易写错、最该每个平台都跑到的一条。

选择逻辑本身与任何 Win32 调用无关，挪进不带 SupportedOSPlatform 的 ScreenTargetPicker。
留在 WindowsScreenCapture 里时它会继承那个类的平台特性，测试跟着只能在 Windows 上跑；
挪出来之后三个平台都真跑。假 Win32 层随之删掉，用例直接构造候选窗口。

另删掉「平台不支持时报告不可用」：它在非 Windows 上构造 Windows 专用类型，本身就违反平台
约束（CA1416），而它要表达的后果已由 ScreenToolsTests.平台不支持截屏时不注册 覆盖。

变异验证重做：去掉「排除本进程」之后三条用例失败。
